requirement.txt
streamlit
seaborn
pandas
matplotlib.pyplot
plotly.express